requirements.txt
fastapi
uvicorn
pymongo[srv]
passlib[bcrypt]
pydantic
python-dotenv
motor
pydantic[email]
email-validator
python-jose[cryptography]
bcrypt